Add paging file in mem module
