+++ src/sys/mem/heap.rs
@@ -3,8 +3,7 @@ use crate::sys;
 use core::cmp;
 use linked_list_allocator::LockedHeap;
 use x86_64::structures::paging::{
-    mapper::MapToError, page::PageRangeInclusive,
-    FrameAllocator, Mapper, OffsetPageTable, Page, PageTableFlags, Size4KiB,
+    mapper::MapToError, FrameAllocator, Mapper, Page, PageTableFlags, Size4KiB
 };
 use x86_64::VirtAddr;
 
@@ -50,65 +49,6 @@ pub fn init_heap() -> Result<(), MapToError<Size4KiB>> {
     }
 
     Ok(())
-}
-
-pub fn alloc_pages(
-    mapper: &mut OffsetPageTable, addr: u64, size: usize
-) -> Result<(), ()> {
-    let size = size.saturating_sub(1) as u64;
-    let mut frame_allocator = sys::mem::frame_allocator();
-
-    let pages = {
-        let start_page = Page::containing_address(VirtAddr::new(addr));
-        let end_page = Page::containing_address(VirtAddr::new(addr + size));
-        Page::range_inclusive(start_page, end_page)
-    };
-
-    let flags = PageTableFlags::PRESENT
-              | PageTableFlags::WRITABLE
-              | PageTableFlags::USER_ACCESSIBLE;
-
-    for page in pages {
-        if let Some(frame) = frame_allocator.allocate_frame() {
-            let res = unsafe {
-                mapper.map_to(page, frame, flags, &mut frame_allocator)
-            };
-            if let Ok(mapping) = res {
-                //debug!("Mapped {:?} to {:?}", page, frame);
-                mapping.flush();
-            } else {
-                debug!("Could not map {:?} to {:?}", page, frame);
-                if let Ok(old_frame) = mapper.translate_page(page) {
-                    debug!("Already mapped to {:?}", old_frame);
-                }
-                return Err(());
-            }
-        } else {
-            debug!("Could not allocate frame for {:?}", page);
-            return Err(());
-        }
-    }
-
-    Ok(())
-}
-
-// TODO: Replace `free` by `dealloc`
-pub fn free_pages(mapper: &mut OffsetPageTable, addr: u64, size: usize) {
-    let size = size.saturating_sub(1) as u64;
-
-    let pages: PageRangeInclusive<Size4KiB> = {
-        let start_page = Page::containing_address(VirtAddr::new(addr));
-        let end_page = Page::containing_address(VirtAddr::new(addr + size));
-        Page::range_inclusive(start_page, end_page)
-    };
-
-    for page in pages {
-        if let Ok((_, mapping)) = mapper.unmap(page) {
-            mapping.flush();
-        } else {
-            //debug!("Could not unmap {:?}", page);
-        }
-    }
 }
 
 pub fn heap_size() -> usize {
